put utils func into new file, added ideas to improve performance
