+++ src/search/domain_abstractions/domain_abstraction.h
@@ -25,7 +25,7 @@ namespace domain_abstractions {
 
         VariableGroupVectors variableGroupVectors; // group mapping
         std::vector<long long> nValuesForHash; // NValues need to compute perfect hash function for h-value lookup
-        long long numAbstractStates{};
+        long long numAbstractStates{0};
 
         // Operators based on abstraction(group var)
         std::vector<std::vector<std::pair<int, int>>> abstractOperatorPreconditions;
@@ -55,8 +55,6 @@ namespace domain_abstractions {
 
         std::vector<FactPair> getVariableGroupFacts(int varIndex, int groupNumber);
 
-        long long int getNumAbstractStates() const;
-
         std::vector<std::shared_ptr<DomainAbstractedState>> getAbstractGoalStates();
 
         DomainAbstractedStates getPredecessors(const std::shared_ptr<DomainAbstractedState> &state);
@@ -68,42 +66,6 @@ namespace domain_abstractions {
     private:
 
         std::vector<long long> computeNValues(VariableGroupVectors newAbstraction);
-
-        std::vector<std::vector<int>> groupCombinations(std::vector<std::vector<int>> groupNumbersPerVar) {
-            // Init
-            int numResultingStates = 1;
-            std::vector<std::vector<int>::iterator> iterators;
-            assert(groupNumbersPerVar.size() == originalTask.get_variables().size());
-
-            int howManyVectors = (int) groupNumbersPerVar.size();
-            for (auto & i : groupNumbersPerVar) {
-                numResultingStates *= (int) i.size();
-                iterators.push_back(i.begin());
-            }
-
-            assert(iterators.size() == groupNumbersPerVar.size());
-            assert(howManyVectors == (int) groupNumbersPerVar.size());
-            std::vector<std::vector<int>> resultingStates;
-            resultingStates.reserve(numResultingStates);
-
-            // Odometer to generate combinations
-            while (iterators.at(0) != groupNumbersPerVar.at(0).end()) {
-                // process the pointed-to elements
-                std::vector<int> toInsert;
-                for (int i = 0; i < howManyVectors; i++) {
-                    toInsert.push_back(*(iterators[i]));
-                }
-                assert(toInsert.size() == groupNumbersPerVar.size());
-                resultingStates.push_back(toInsert);
-                // the following increments the "odometer" by 1
-                ++iterators[howManyVectors - 1];
-                for (int i = howManyVectors - 1; (i > 0) && (iterators[i] == groupNumbersPerVar[i].end()); --i) {
-                    iterators[i] = groupNumbersPerVar[i].begin();
-                    ++iterators[i - 1];
-                }
-            }
-            return resultingStates;
-        };
 
         static bool
         abstractStateFulfillsAbstractFacts(std::vector<int> abstractState, const std::vector<std::pair<int, int>> &abstractFacts);
